requirements.txt
google-genai 
streamlit 
pillow 
